phase 8

+++ doc/phase8/volatile/read/expected.c
@@ -5,6 +5,6 @@ volatile uint8_t reg = 0xFF;
 int main(void) {
     TSC_INIT();
     const uint8_t v = reg;
-    printf("%u\n", v);
+    printf("%u\n", (unsigned)v);
     return 0;
 }
